Split local parallelism helpers into package
