+++ mhcflurry/cli/select_processing_models_command.py
@@ -37,7 +37,7 @@ import tqdm  # progress bar
 from ..class1_processing_predictor import Class1ProcessingPredictor
 from ..flanking_encoding import FlankingEncoding
 from ..common import configure_logging, write_generate_sh
-from ..local_parallelism import (
+from ..parallelism import (
     resolve_local_parallelism_args,
     worker_pool_with_gpu_assignments_from_args,
     add_local_parallelism_args)
