+++ mhcflurry/cli/train_pan_allele_models_command.py
@@ -47,7 +47,7 @@ from ..common import (
     normalize_allele_name,
     write_generate_sh,
 )
-from ..local_parallelism import (
+from ..parallelism import (
     add_local_parallelism_args,
     apply_resolved_training_hyperparameters_to_work_items,
     attach_constant_data_to_work_items_if_needed,
