+++ mhcflurry/cli/train_processing_models_command.py
@@ -45,7 +45,7 @@ from ..common import (
     derive_seed,
     write_generate_sh,
 )
-from ..local_parallelism import (
+from ..parallelism import (
     add_local_parallelism_args,
     attach_constant_data_to_work_items_if_needed,
     resolve_local_parallelism_args,
